requirements.txt
tensorflow>=2.12.0
streamlit>=1.32.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0